spice-webdavd: run service in a loop
Although on Linux, the service can be restarted cleanly by systemd when
needed, it's not the case on Windows. And, I can't see a good way to
wait for the virtio channel to be openend.

To solve those 2 related issues, let's start the service and read from
the channel. If an error is returned, wait a second and try again.  That
way, the service can be started on Windows even when no clients are
connected.