spice-webdavd-windows: Automount shared folder
Try to connect to shared folder automatically on Windows.

On each loop of run_service(), run a GTask, that waits for half a
second.  If read_thread() returns, it means, that sharing is not yet
connected, and the map_drive task is cancelled.

If the map_drive task is NOT cancelled, it enumerates free drive
letters, and maps Spice Folder to highest possible (from Z: to A:).

If all drive letters are already assigned, show a critical error, but
don't stop the service. The folder can still be accessed trough other
means.

Acked-by: Victor Toso <victortoso@redhat.com>