spice: refcount Client
A pointer to a Client struct is passed as user_data to async
operations. Count the references to it to avoid memory errors.

Example situation:
Spice-gtk disconnects, the cancellable is cancelled, the main
loop quits and all Clients are removed from the hash table and
freed.
Once another spice-gtk connection is established and the loop
runs again, client_read_cb() for a freed client with
G_IO_ERROR_CLOSED is called.

Signed-off-by: Jakub Janků <jjanku@redhat.com>