spice: OutputQueue refactor
The main issue with the current implementation is that
g_output_stream_write_all() in output_queue_idle()
is called synchronously in the main context.
This is fine for GBufferedOutputStream, since the data is just
copied to a buffer, but not for GUnixOutputStream or
GWin32OutputStream (class of mux_ostream) which might block
the main thread.

OutputQueue is also used for each client, where it doesn't make
much sense because the queue has never more than 1 item at a time.
So let's just use g_output_stream_write_all() instead.

Signed-off-by: Jakub Janků <jjanku@redhat.com>