propfind: calculate quota for the appropriate resource
Currently, available quota is calculated for the root.
However, if the resource resides e.g. on a different partition
because of symlinks, the returned value is incorrect.

According to [0], the calculation for used quota might be left
as it is, but let's change it as well, for the sake of consistency.

[0] http://www.webdav.org/specs/rfc4331.html#quota-used-bytes

Signed-off-by: Jakub Janků <jjanku@redhat.com>