libphodav: rename Lock* enums
The name 'LOCK_WRITE' is already taken by Linux. If a system header
including 'asm-generic/fcntl.h' is included, it causes compilation error
because the '#define' in the Linux header makes it a syntax error. This
is known to happen on Android. To avoid the problem, rename all Lock*
types to DAVLock*.

Fixes: https://gitlab.gnome.org/GNOME/phodav/issues/2