server+ci: test PhodavVirtualDir
Add new project subdir "tests".

Add "tests/virtual-dir-server.c". This is in principle similar
to chezdav but uses the new PhodavVirtualDir functions.

Test virtual-dir-server using litmus in Gitlab CI
to show that the behavior of phodav is not affected when using
the "virtual root".

Also test virtual-dir-server using "tests/virtual-dir.c" which
aims to cover the parts litmus cannot. This is quite limited
and only checks that the correct status is returned.

Signed-off-by: Jakub Janků <jjanku@redhat.com>