server: don't overwrite error in delete enumeration
If g_file_enumerator_next_file() returns NULL and sets error,
the error variable must be cleared before it can be used
with g_file_delete() beneath the while loop.

Otherwise, if g_file_delete() fails as well, the error is
overwritten and GLib complains:

"GError set over the top of a previous GError or uninitialized memory.
This indicates a bug in someone's code.
You must ensure an error is NULL before it's set.
The overwriting error message was: %s"

Signed-off-by: Jakub Janků <jjanku@redhat.com>