tests: launch virtual-dir-server using GSubprocess
Use g_subprocess_new to spawn the test server.
This way the tests can also be used on Windows.

If the parent process with the tests aborts, we don't want to
leave the test server running. Use G_SUBPROCESS_FLAGS_STDIN_PIPE
to setup a pipe to the test server and add an io channel watch
in the server. If the channel becomes ready for reading, quit.

Fixes: https://gitlab.gnome.org/GNOME/phodav/-/issues/14

Signed-off-by: Jakub Janků <jjanku@redhat.com>